Fix calculation of notification and event date
0fc91a3149e9b550e802948eddcc13ca795012d5 tried to shift the meeting by a
week, with an interesting[tm] result:

```
./mail-reminder
./data/board-meeting.toml: event on 2025-04-07, reminder to be sent on 2025-04-13.
```

That would send the reminder at the right date, but for a meeting at a
wrong date in the past - with the result that no reminders were sent.
(The actual reason was that on the day the reminder would be sent, the
 script already calculated the date of the next meeting, with the
 reminder to be sent in about two weeks.)

Fix this by reverting the "Shift by 1 week" code changes, and instead
adjust biweekly_week in the toml file to match the correct meeting date.

With that, the script output looks more promising:

```
./mail-reminder
./data/board-meeting.toml: event on 2025-04-14, reminder to be sent on 2025-04-13.
```